+++ src/Controller/DisponibiliteController.php
@@ -11,6 +11,7 @@ use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Security\Http\Attribute\IsGranted;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
 
 #[Route('/api/disponibilites', name: 'api_disponibilites_')]
@@ -42,6 +43,7 @@ class DisponibiliteController extends AbstractController
     }
 
     #[Route('', name: 'create', methods: ['POST'])]
+    #[IsGranted('ROLE_PROPRIETAIRE')]
     public function create(Request $request): JsonResponse
     {
         $data = json_decode($request->getContent(), true);
@@ -60,6 +62,10 @@ class DisponibiliteController extends AbstractController
 
         if (!$bateau) {
             return $this->json(['message' => 'Bateau introuvable.'], Response::HTTP_UNPROCESSABLE_ENTITY);
+        }
+
+        if (!$this->isGranted('ROLE_ADMIN') && $bateau->getProprietaire() !== $this->getUser()) {
+            return $this->json(['message' => 'Accès refusé.'], Response::HTTP_FORBIDDEN);
         }
 
         $dispo = new Disponibilite();
@@ -88,6 +94,10 @@ class DisponibiliteController extends AbstractController
             return $this->json(['message' => 'Disponibilité non trouvée.'], Response::HTTP_NOT_FOUND);
         }
 
+        if (!$this->isGranted('ROLE_ADMIN') && $dispo->getBateau()->getProprietaire() !== $this->getUser()) {
+            return $this->json(['message' => 'Accès refusé.'], Response::HTTP_FORBIDDEN);
+        }
+
         $data = json_decode($request->getContent(), true);
 
         if (isset($data['date_debut'])) $dispo->setDateDebut(new \DateTime($data['date_debut']));
@@ -106,6 +116,10 @@ class DisponibiliteController extends AbstractController
 
         if (!$dispo) {
             return $this->json(['message' => 'Disponibilité non trouvée.'], Response::HTTP_NOT_FOUND);
+        }
+
+        if (!$this->isGranted('ROLE_ADMIN') && $dispo->getBateau()->getProprietaire() !== $this->getUser()) {
+            return $this->json(['message' => 'Accès refusé.'], Response::HTTP_FORBIDDEN);
         }
 
         $this->em->remove($dispo);
